Display and add patients
Contains several adjustments to support adding new patient, and displaying them in a table.

The fake db indexed memory object containing patients can now be recomputed from the source fake database.

The CSS had not been hooked up yet, so it was added as SCSS as webpack supported that already.